requirements.txt
pyswmm>=1.4.0
swmm-toolkit>=0.9.0
networkx>=2.8
pandas>=1.5.0
numpy>=1.21.0
xgboost>=1.7.0
lightgbm>=3.3.0
scikit-learn>=1.2.0
torch>=1.13.0
torch-geometric>=2.3.0
mlflow>=2.0.0
onnxruntime>=1.14.0
onnxmltools>=1.11.0
skl2onnx>=1.15.0
joblib>=1.2.0
pyyaml>=6.0</content>
<parameter name="filePath">c:\Users\kamlo\Desktop\Personal\projects\swmm_project\requirements.txt